+++ src/main/java/org/manuel/mysportfolio/constants/Constants.java
@@ -1,0 +1,10 @@
+package org.manuel.mysportfolio.constants;
+
+public class Constants {
+
+  public static final int COMPETITION_NAME_MAX_CHARACTERS = 40;
+  public static final int TEAM_NAME_MAX_CHARACTERS = 40;
+
+  public static final int MATCH_DESCRIPTION_MAX_CHARACTERS = 400;
+
+}