+++ src/main/java/org/manuel/mysportfolio/controllers/query/user/UserPlayerProfileQueryController.java
@@ -6,8 +6,10 @@ import org.manuel.mysportfolio.model.entities.user.AppUser;
 import org.manuel.mysportfolio.services.query.AppUserQueryService;
 import org.manuel.mysportfolio.services.query.PlayerProfileQueryService;
 import org.manuel.mysportfolio.transformers.PlayerProfileToPlayerProfileDtoTransformer;
+import org.manuel.mysportfolio.validations.UserExists;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -15,6 +17,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/api/v1/users/{userId}/player")
+@Validated
 @lombok.AllArgsConstructor
 public class UserPlayerProfileQueryController {
 
@@ -23,7 +26,7 @@ public class UserPlayerProfileQueryController {
   private final PlayerProfileToPlayerProfileDtoTransformer playerProfileToPlayerProfileDtoTransformer;
 
   @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-  public ResponseEntity<PlayerProfileDto> getPlayer(@PathVariable final String userId) {
+  public ResponseEntity<PlayerProfileDto> getPlayer(@PathVariable @UserExists final String userId) {
     appUserQueryService.findByExternalId(userId)
         .orElseThrow(() -> new EntityNotFoundException(AppUser.class, userId));
     return ResponseEntity.ok(playerProfileToPlayerProfileDtoTransformer