+++ src/main/java/org/manuel/mysportfolio/model/dtos/user/UserTeamDto.java
@@ -4,6 +4,8 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
 import io.github.manuelarte.spring.manuelartevalidation.constraints.groups.New;
+import io.github.manuelarte.spring.manuelartevalidation.constraints.groups.Update;
+import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
 import org.manuel.mysportfolio.model.dtos.team.TeamDto;
 import org.manuel.mysportfolio.model.dtos.teamtousers.UserInTeamDto;
@@ -15,9 +17,12 @@ import org.manuel.mysportfolio.model.dtos.teamtousers.UserInTeamDto;
 @lombok.Builder(toBuilder = true)
 public class UserTeamDto {
 
-  @NotNull(groups = {New.class})
+  @NotNull(groups = {New.class, Update.class})
+  @Valid
   private final TeamDto team;
 
+  @NotNull(groups = {New.class, Update.class})
+  @Valid
   private final UserInTeamDto userInTeam;
 
   @JsonPOJOBuilder(withPrefix = "")