+++ src/main/java/org/manuel/mysportfolio/model/entities/match/AnonymousTeam.java
@@ -2,6 +2,7 @@ package org.manuel.mysportfolio.model.entities.match;
 
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.Size;
+import org.manuel.mysportfolio.constants.Constants;
 import org.manuel.mysportfolio.model.TeamInfo;
 import org.manuel.mysportfolio.model.entities.team.KitPart;
 import org.manuel.mysportfolio.model.entities.team.TeamImage;
@@ -13,7 +14,7 @@ import org.manuel.mysportfolio.model.entities.team.TeamKit;
 public class AnonymousTeam implements TeamType, TeamInfo {
 
   @NotEmpty
-  @Size(max = 30)
+  @Size(max = Constants.TEAM_NAME_MAX_CHARACTERS)
   private String name;
 
   private TeamKit<KitPart, KitPart> teamKit;