+++ src/main/java/org/manuel/mysportfolio/publishers/TeamCreatedEventPublisher.java
@@ -10,9 +10,10 @@ import org.springframework.stereotype.Component;
 @lombok.extern.slf4j.Slf4j
 public class TeamCreatedEventPublisher {
 
-  private ApplicationEventPublisher applicationEventPublisher;
+  private final ApplicationEventPublisher applicationEventPublisher;
 
   public void publishEvent(final Team createdTeam) {
+    assert createdTeam.getId() != null;
     log.info("Publishing team created event for team {}.", createdTeam.getId().toString());
     final var teamCreatedEvent = new TeamCreatedEvent(createdTeam);
     applicationEventPublisher.publishEvent(teamCreatedEvent);