+++ CHANGELOG.md
@@ -8,7 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Adding @UserExists annotation
+- Adding PUT endpoint in UserTeams
+
 ### Changed
+
+- Increasing the max number of characters for team name, competition name and description and match description
 
 ### Removed